Store API: extract checkout place-order pipeline into ProcessCheckoutTrait (pure move)
Move the web Checkout route's place-order orchestration — request lifecycle
(GET/POST/PUT), draft-order creation, customer processing, pre-payment validation,
stock/coupon holds and payment — plus requires_nonce, verbatim into a new @internal
ProcessCheckoutTrait, and compose it on the route.

Pure move: no content or behaviour change. Verify with `git show --color-moved=zebra`.
Lint/PHPStan adjustments for the moved code and the changelog follow in the next commit.
